// src/common/dto/parameter.dto.ts
// {  "idParametros": "string",  "nombreParametro": "string",  "tipoDato": "string",  "estado": "string",  "tipoValor": "string",  "valorDesde": "string",  "valorHasta": "string",  "listaValores": "string",  "idGrupo": "string"}

export class ParameterDto {
    idParametros: string;
    nombreParametro: string;
    tipoDato: string;
    estado: string;
    tipoValor: string;
    valorDesde: string;
    valorHasta: string;
    listaValores: string;
    idGrupo: string;
}